[#188] inputPort, OutputPort 정의

+++ Architecture/Book-rental-service/src/main/java/io/chan/bookrentalservice/framework/web/dto/ClearOverdueInfoDto.java
@@ -3,6 +3,6 @@ package io.chan.bookrentalservice.framework.web.dto;
 public record ClearOverdueInfoDto(
     String userId,
     String username,
-    int point
+    Long point
 ) {
 }
